requirements.txt
Django==4.1.3
psycopg2-binary==2.9.5
djangorestframework==3.14.0
python-dotenv==0.21.0
